refactor: move provider initialization into factory
